fix: resolve NUMERIC to f64 type mismatch in username stats
Problem: PostgreSQL returns NUMERIC type for percentage calculation,
but Rust expects f64 (FLOAT8)

Solution: Added ::float8 cast to SQL query

Before:
(COUNT(*) * 100.0 / NULLIF($1, 0)) as percentage  -- NUMERIC type

After:
(COUNT(*) * 100.0 / NULLIF($1, 0))::float8 as percentage  -- FLOAT8 type

Dashboard command now works correctly! ✅

Test result:
- Total Users: 21,651
- Total Activities: 334,104
- Total Casts: 87,627
- Stats display properly formatted

+++ src/database.rs
@@ -679,7 +679,7 @@ impl Database {
             SELECT 
                 up.username,
                 COUNT(DISTINCT uat.id) as count,
-                (COUNT(DISTINCT uat.id) * 100.0 / NULLIF($1, 0)) as percentage
+                (COUNT(DISTINCT uat.id) * 100.0 / NULLIF($1, 0))::float8 as percentage
             FROM user_profiles up
             LEFT JOIN user_activity_timeline uat ON up.fid = uat.fid
             WHERE up.username IS NOT NULL AND up.username != ''
